requirements.txt
apache-airflow-providers-dbt-cloud==4.3.2
apache-airflow-providers-papermill==3.11.3
pyspark==3.5.7
apache-airflow-providers-apache-spark==5.3.2